代码整理；分菜单栏使用stateman实现

+++ zuiwan/controllers/Media.php
@@ -50,8 +50,9 @@ class Media extends MY_Controller
                         $media = $this->media->get_media_by_name($media_name);
                         $origin = $media['media_avatar'];
                         $origin_pos = STATIC_PATH . $this->config->config['img_dir'] . "/" . $origin;
-                        unlink($origin_pos);
-
+                        if (file_exists($origin_pos)){
+                            unlink($origin_pos);
+                        }
                         //把media的media_avatar更新
                         $this->media->update_media_avatar($media_name, $random_file_name);
                         $result['data'] = $random_file_name;
